Moved TaskHolder and Adapter to a separate file
This is needed for the subtask view in the detail view.
