+++ app/src/main/java/com/task4/task4/TaskListFragment.kt
@@ -4,14 +4,11 @@ import android.content.Context
 import android.os.Bundle
 import android.util.Log
 import android.view.*
-import android.widget.CheckBox
-import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.task4.task4.database.Task
-import java.util.UUID
 
 private const val TAG = "TaskListFragment"
 
@@ -21,64 +18,15 @@ class TaskListFragment : Fragment() {
     private lateinit var taskRecyclerView: RecyclerView
 
     // adapts the list of tasks  from the database for the recycler view
-    private var adapter: TaskAdapter? = TaskAdapter(emptyList())
+    private var adapter = TaskAdapter(emptyList())
 
     // preserve data across rotations lazy load it in
     private val taskListViewModel: TaskListViewModel by viewModels()
 
-    private inner class TaskHolder(view: View) : RecyclerView.ViewHolder(view),
-                                                 View.OnClickListener {
-
-        private lateinit var task: Task
-        private val titleTextView: TextView = itemView.findViewById(R.id.task_title)
-        private val dateTextView: TextView = itemView.findViewById(R.id.task_due_date)
-        private val doneCheckBox: CheckBox = itemView.findViewById(R.id.task_completed)
-
-        fun bind(task: Task) {
-            this.task = task
-            this.task.apply {
-                titleTextView.text = name
-                dateTextView.text = userDate
-                doneCheckBox.apply {
-                    isChecked = completed
-                    jumpDrawablesToCurrentState()
-                }
-            }
-        }
-
-        init {
-            itemView.setOnClickListener(this)
-        }
-
-        override fun onClick(v: View?) {
-            callbacks?.onTaskSelected(task.id)
-        }
-    }
-
-    interface Callbacks {
-
-        fun onTaskSelected(taskId: UUID)
-    }
-
-    private var callbacks: Callbacks? = null
-
-    private inner class TaskAdapter(var tasks: List<Task>) : RecyclerView.Adapter<TaskHolder>() {
-
-        override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): TaskHolder = TaskHolder(
-            layoutInflater.inflate(R.layout.list_item_task, parent, false)
-        )
-
-        override fun getItemCount() = tasks.size
-
-        override fun onBindViewHolder(holder: TaskHolder, position: Int) {
-            holder.bind(tasks[position])
-        }
-
-    }
-
     override fun onAttach(context: Context) {
         super.onAttach(context)
-        callbacks = context as Callbacks?
+        adapter.layoutInflater = layoutInflater
+        adapter.callbacks = context as TaskRecylerViewCallbacks?
     }
 
     override fun onCreateView(
@@ -86,27 +34,27 @@ class TaskListFragment : Fragment() {
     ): View? {
         val view = inflater.inflate(R.layout.fragment_task_list, container, false)
         taskRecyclerView = view.findViewById(R.id.task_recycler_view)
-
-        taskRecyclerView.layoutManager = LinearLayoutManager(context)
-        taskRecyclerView.adapter = adapter
+        taskRecyclerView.apply {
+            layoutManager = LinearLayoutManager(context)
+            adapter = adapter
+        }
         return view
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        taskListViewModel.taskListLiveData.observe(viewLifecycleOwner, { tasks ->
+        taskListViewModel.taskListLiveData.observe(viewLifecycleOwner) { tasks ->
             tasks?.let {
                 Log.i(TAG, "Got tasks ${tasks.size}")
                 updateUI(tasks)
             }
-
-        })
+        }
     }
 
     // remove the reference to the callback receiver
     override fun onDetach() {
         super.onDetach()
-        callbacks = null
+        adapter.callbacks = null
     }
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -123,15 +71,14 @@ class TaskListFragment : Fragment() {
         R.id.new_task -> {
             val task = Task()
             taskListViewModel.addTask(task)
-            callbacks?.onTaskSelected(task.id)
+            adapter.callbacks?.onTaskSelected(task.id)
             true
         }
         else          -> super.onOptionsItemSelected(item)
     }
 
-
     private fun updateUI(tasks: List<Task>) {
-        adapter = TaskAdapter(tasks)
+        adapter.tasks = tasks
         taskRecyclerView.adapter = adapter
     }
